requirements.txt
streamlit
transformers
torch
sentencepiece  
pdfplumber    
pandas         
Pillow       
